\name{cox_pls2_supp_vals_KM}
\alias{cox_pls2_supp_vals_KM}
\docType{data}
\title{
Demo dataset
}
\description{
Dataset used by one of the demos
}
\usage{data(cox_pls2_supp_vals_KM)}
\format{
  The format is:
 num [1:56] 0.899 0.899 0.734 0.737 0.698 ...
}
\references{plsRcox : \enc{modèles}{mod\`eles} de Cox en \enc{présence}{pr\'esence} d'un grand nombre de variables explicatives, \enc{Frédéric}{Fr\'ed\'eric} Bertrand, Myriam Maumy-Bertrand, Marie-Pierre Gaub, Nicolas Meyer, \enc{Chimiométrie}{Chimiom\'etrie} 2010, Paris, 2010.}
\examples{
data(cox_pls2_supp_vals_KM)
}
\keyword{datasets}
