\name{lars_supp_vals_NNE}
\alias{lars_supp_vals_NNE}
\docType{data}
\title{
Demo dataset
}
\description{
Dataset used by one of the demos
}
\usage{data(lars_supp_vals_NNE)}
\format{
  The format is:
 num [1:56] 0.888 0.888 0.651 0.571 0.539 ...
}
\references{plsRcox : \enc{modèles}{mod\`eles} de Cox en \enc{présence}{pr\'esence} d'un grand nombre de variables explicatives, \enc{Frédéric}{Fr\'ed\'eric} Bertrand, Myriam Maumy-Bertrand, Marie-Pierre Gaub, Nicolas Meyer, \enc{Chimiométrie}{Chimiom\'etrie} 2010, Paris, 2010.}
\examples{
data(lars_supp_vals_NNE)
}
\keyword{datasets}
