\name{lars_supp_vals_NNE_micro}
\alias{lars_supp_vals_NNE_micro}
\docType{data}
\title{
Demo dataset
}
\description{
Dataset used by one of the demos
}
\usage{data(lars_supp_vals_NNE_micro)}
\format{
  The format is:
 num [1:27] 0.698 0.698 0.698 0.698 0.698 ...
}
\references{plsRcox : \enc{modèles}{mod\`eles} de Cox en \enc{présence}{pr\'esence} d'un grand nombre de variables explicatives, \enc{Frédéric}{Fr\'ed\'eric} Bertrand, Myriam Maumy-Bertrand, Marie-Pierre Gaub, Nicolas Meyer, \enc{Chimiométrie}{Chimiom\'etrie} 2010, Paris, 2010.}
\examples{
data(lars_supp_vals_NNE_micro)
}
\keyword{datasets}
