% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-plsRcox.R
\name{internal-plsRcox}
\alias{internal-plsRcox}
\alias{getIndic}
\alias{getIndicCV}
\alias{getIndicCViAUCSH}
\alias{getIndicCViAUCSurvROCTest}
\alias{correctp.cox}
\alias{spls.cox}
\alias{ust}
\alias{spls.dv}
\alias{pls.cox}
\alias{predict.pls.cox}
\title{Internal plsRcox functions}
\description{
These are not to be called by the user.
}
\references{
plsRcox, Cox-Models in a high dimensional setting in R, Frederic
Bertrand, Philippe Bastien, Nicolas Meyer and Myriam Maumy-Bertrand (2014).
Proceedings of User2014!, Los Angeles, page 152.\cr

Deviance residuals-based sparse PLS and sparse kernel PLS regression for
censored data, Philippe Bastien, Frederic Bertrand, Nicolas Meyer and Myriam
Maumy-Bertrand (2015), Bioinformatics, 31(3):397-404,
doi:10.1093/bioinformatics/btu660.
}
\author{
Frédéric Bertrand\cr
\email{frederic.bertrand@lecnam.net}\cr
\url{https://fbertran.github.io/homepage/}
}
\keyword{internal}
